refactor: several adaptions
