+++ streampipes-data-explorer/src/main/java/org/apache/streampipes/dataexplorer/AutoAggregationHandler.java
@@ -15,15 +15,14 @@
  * limitations under the License.
  *
  */
-package org.apache.streampipes.dataexplorer.influx.migrate;
+package org.apache.streampipes.dataexplorer;
 
-import org.apache.streampipes.dataexplorer.DataExplorerSchemaManagement;
+import org.apache.streampipes.dataexplorer.api.IDataExplorerQueryManagement;
 import org.apache.streampipes.model.datalake.param.ProvidedRestQueryParams;
 import org.apache.streampipes.model.datalake.param.SupportedRestQueryParams;
 import org.apache.streampipes.dataexplorer.param.model.SelectColumn;
 import org.apache.streampipes.model.datalake.DataLakeQueryOrdering;
 import org.apache.streampipes.model.datalake.SpQueryResult;
-import org.apache.streampipes.storage.management.StorageDispatcher;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -50,16 +49,9 @@ public class AutoAggregationHandler {
   private final IDataExplorerQueryManagement dataLakeQueryManagement;
   private final ProvidedRestQueryParams queryParams;
 
-  public AutoAggregationHandler(ProvidedRestQueryParams params) {
+  public AutoAggregationHandler(ProvidedRestQueryParams params, IDataExplorerQueryManagement dataExplorerQueryManagement) {
     this.queryParams = params;
-    this.dataLakeQueryManagement = getDataLakeQueryManagement();
-  }
-
-  private IDataExplorerQueryManagement getDataLakeQueryManagement() {
-    var dataLakeStorage = StorageDispatcher.INSTANCE
-        .getNoSqlStore()
-        .getDataLakeStorage();
-    return new DataExplorerQueryManagement(new DataExplorerSchemaManagement(dataLakeStorage));
+    this.dataLakeQueryManagement = dataExplorerQueryManagement;
   }
 
   public ProvidedRestQueryParams makeAutoAggregationQueryParams() throws IllegalArgumentException {
