+++ streampipes-data-explorer/src/main/java/org/apache/streampipes/dataexplorer/ImageStoreUtils.java
@@ -30,7 +30,8 @@ public class ImageStoreUtils {
   public static List<EventProperty> getImageProperties(DataLakeMeasure measure) {
     return measure.getEventSchema().getEventProperties().stream()
         .filter(eventProperty -> eventProperty.getDomainProperties() != null
-            && eventProperty.getDomainProperties().size() > 0
+            && !eventProperty.getDomainProperties()
+                             .isEmpty()
             && eventProperty.getDomainProperties().get(0).toString().equals(SPSensor.IMAGE))
         .collect(Collectors.toList());
   }
