+++ streampipes-data-explorer/src/main/java/org/apache/streampipes/dataexplorer/TimeSeriesStore.java
@@ -39,7 +39,6 @@ public class TimeSeriesStore {
 
   public TimeSeriesStore(
     ITimeSeriesStorage timeSeriesStorage,
-    IStreamPipesClient client,
     DataLakeMeasure measure,
     Environment environment,
     boolean enableImageStore
